show weather and forecast data

+++ app/src/main/java/com/mahedi/weather_app_0400034063/models/CurrentWeatherModel.kt
@@ -12,7 +12,7 @@ data class CurrentModel(
     @SerializedName("coord")
     val coord: Coord,
     @SerializedName("dt")
-    val dt: Int,
+    val dt: Long,
     @SerializedName("id")
     val id: Int,
     @SerializedName("main")
@@ -63,9 +63,9 @@ data class CurrentModel(
         @SerializedName("id")
         val id: Int,
         @SerializedName("sunrise")
-        val sunrise: Int,
+        val sunrise: Long,
         @SerializedName("sunset")
-        val sunset: Int,
+        val sunset: Long,
         @SerializedName("type")
         val type: Int
     )
